requirements.txt
bs4>=0.0.1
python-dotenv>=1.0.0
openai>=0.27.0
graphviz>=0.20.1